feat: removeliquidity and lp balance
