+++ src/plugins/foxPage/const.ts
@@ -6,4 +6,3 @@ export const UNISWAP_V2_WETH_FOX_FARMING_REWARDS_ADDRESS =
 export const UNISWAP_V4_WETH_FOX_FARMING_REWARDS_ADDRESS =
   '0x24fd7fb95dc742e23dc3829d3e656feeb5f67fa0'
 export const TRADING_FEE_RATE = 0.003
-export const UNISWAP_V2_ROUTER_ADDRESS = '0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D'
